test date

+++ post-to-vk.php
@@ -68,17 +68,19 @@ function getSavedAttachments($post_attachments){
 
 function deleteSavedAttachments(){
     $dir = 'images'. DIRECTORY_SEPARATOR . 'posts'. DIRECTORY_SEPARATOR .'post100';
-    $it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
-    $files = new RecursiveIteratorIterator($it,
-        RecursiveIteratorIterator::CHILD_FIRST);
-    foreach($files as $file) {
-        if ($file->isDir()){
-            rmdir($file->getRealPath());
-        } else {
-            unlink($file->getRealPath());
+    if(is_dir($dir)){
+        $it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
+        $files = new RecursiveIteratorIterator($it,
+            RecursiveIteratorIterator::CHILD_FIRST);
+        foreach($files as $file) {
+            if ($file->isDir()){
+                rmdir($file->getRealPath());
+            } else {
+                unlink($file->getRealPath());
+            }
         }
+        rmdir($dir);
     }
-    rmdir($dir);
 }
 
 function getReplacedPostText($post_text){
